+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 # Changelog
 
+## TODO - x.x.x
+
+- [mu miror](https://editor.p5js.org/wyattroy/sketches/s1v8O0WBf) from [youtube](https://www.youtube.com/watch?v=2BCMZMRFljI&t=1s) in combination with music visualizer
+
+## 0.6.0
+
+- Music Visualizer (Alpha)
+
 ## 0.5.0
 
 - reading list for 2025 started
@@ -27,4 +35,3 @@
 ## 0.1.0
 
 - Initial landing page with GitHub and Instagram links.
-